Rename column names
When faced with column names that have capital letters or space in between, very often we need to change the column names and replace them with lowercase and underscore — to be more clean and Pythonic. Also, we want to make the column names as explicit as possible such that your friends will roughly know what a column contains just by looking at the column name.