List comprehension
You may think that list comprehension is so common and wonder how that could be considered as part of data cleaning. Well… To me, list comprehension is so elegant and clean that you can abstract your logic in a single line of code without having any for loops, and compute that much faster than for loops.
Typically, I’d use list comprehension if I want to get a list of values based on certain conditions to append to the existing dataframe or use that for further analysis.