Output df based on unique last strings in a column
Using back the dataframe with all customers’ information and adding a column of timestamp for each id.
Now each id is not unique and it’s repeated throughout their respective time period. For each id you want to get the last row of the id because you only care about the final customers’ information for each id.
This is when you can use the code snippet above to drop all duplicated id and just keep the last row. This makes sure that we always get the final customers’ information with unique id. Again, this can be done with just a single line of code!