Get discrete intervals from numerical values
This is one of my favourite tools when it comes to converting numerical values in a column to discrete intervals based on the range specified. I highly recommend you to use pd.cut if you want to convert a continuous variable to a categorical variable.
For instance, in the code snippet above, we have a rating column that consists of numerical values from 1–10. What if we want to convert these rating values to certain groups within given their values and specified range? We can bin the values in discrete intervals and label them as bad, moderate, good, strong given the range specified using pd.cut.
Another common use case of pd.cut would be to convert ages to groups of age ranges where you can categorize each age to a label.